refactor: Add .dockerignore file to exclude unnecessary files from Docker build
A new .dockerignore file has been added to the repository to specify which files and directories should be excluded from the Docker build process. This helps to reduce the size of the Docker image and improve build times by excluding unnecessary files such as .env, .git, and temporary files generated during development.